Create Ambogo Sees
A real-time face recognition system is capable of identifying or verifying a person from a video frame. To recognize the face in a frame, first, you need to detect whether the face is present in the frame. If it is present, mark it as a region of interest (ROI), extract the ROI and process it for facial recognition.